refactor: extract steps into middleware file
